+++ CHANGELOG.md
@@ -6,7 +6,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+## [0.12.0] - 2024-10-03
+
+### Added
+* add the capacity to interact with user
 
 ### Changed
 * move command name constants in an Enum class